refactor: 도메인 이름 변경 - Reservation, Waiting → ReservationSlot, Reservation
