+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,42 +1,29 @@
 package roomescape.domain;
 
-import java.time.LocalDate;
-
 import roomescape.exception.CustomException;
 import roomescape.exception.ErrorCode;
 
 public class Reservation {
+
     private final Long id;
-    private final LocalDate date;
-    private final Time time;
-    private final Theme theme;
+    private final String name;
+    private final Long reservationId;
 
-    public Reservation(Long id, LocalDate date, Time time, Theme theme) {
-        validateDate(date);
-        validateTime(time);
-        validateTheme(theme);
-
+    public Reservation(Long id, String name, Long reservationId) {
+        validateName(name);
+        
         this.id = id;
-        this.date = date;
-        this.time = time;
-        this.theme = theme;
+        this.name = name;
+        this.reservationId = reservationId;
     }
 
-    private void validateDate(LocalDate date) {
-        if (date == null) {
-            throw new CustomException(ErrorCode.RESERVATION_DATE_NULL);
+    private void validateName(String name) {
+        if (name == null || name.isBlank()) {
+            throw new CustomException(ErrorCode.RESERVATION_NAME_BLANK);
         }
-    }
 
-    private void validateTime(Time time) {
-        if (time == null) {
-            throw new CustomException(ErrorCode.RESERVATION_TIME_NULL);
-        }
-    }
-
-    private void validateTheme(Theme theme) {
-        if (theme == null) {
-            throw new CustomException(ErrorCode.RESERVATION_THEME_NULL);
+        if (name.length() > 255) {
+            throw new CustomException(ErrorCode.RESERVATION_NAME_TOO_LONG);
         }
     }
 
@@ -44,17 +31,11 @@ public class Reservation {
         return id;
     }
 
-
-    public LocalDate getDate() {
-        return date;
+    public String getName() {
+        return name;
     }
 
-    public Time getTime() {
-        return time;
+    public Long getReservationId() {
+        return reservationId;
     }
-
-    public Theme getTheme() {
-        return theme;
-    }
-
 }
