--- src/main/java/roomescape/repository/WaitingDao.java
@@ -1,103 +1,0 @@
-package roomescape.repository;
-
-import java.util.Map;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-
-import roomescape.domain.Waiting;
-
-@Repository
-public class WaitingDao {
-
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert jdbcInsert;
-
-    private final RowMapper<Waiting> waitingRowMapper = (rs, rowNum) -> new Waiting(
-            rs.getLong("id"),
-            rs.getString("name"),
-            rs.getLong("reservation_id")
-    );
-
-    public WaitingDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.jdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("waiting")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    public Long save(String name, Long reservationId) {
-        return jdbcInsert.executeAndReturnKey(Map.of(
-                "name", name,
-                "reservation_id", reservationId
-        )).longValue();
-    }
-
-    public Waiting findById(long id) {
-        String sql = """
-                SELECT  w.id AS wait_id,
-                        w.name AS name,
-                        w.reservation_id AS reservation_id
-                FROM waiting AS w
-                where w.id = ?
-                """;
-        return jdbcTemplate.queryForObject(sql, waitingRowMapper, id);
-    }
-
-    public boolean existByReservationId(long reservationId) {
-        String sql = """
-            SELECT EXISTS(
-                SELECT 1
-                FROM waiting w
-                WHERE w.reservation_id = ?
-            )
-            """;
-
-        return Boolean.TRUE.equals(
-                jdbcTemplate.queryForObject(sql, Boolean.class, reservationId
-                )
-        );
-    }
-
-    public int findOrderByReservationId(long id, long reservationId) {
-        String sql = """
-                SELECT COUNT(*)
-                FROM waiting w
-                WHERE w.reservation_id = ?
-                  AND w.id < ?
-                """;
-        return jdbcTemplate.queryForObject(
-                sql,
-                Integer.class,
-                reservationId,
-                id
-        );
-    }
-
-    public boolean existsByReservation(Long reservationId) {
-        String sql = """
-            SELECT COUNT(*) > 0
-            FROM waiting w
-            WHERE w.reservation_id = ?
-            """;
-        return Boolean.TRUE.equals(jdbcTemplate.queryForObject(sql, Boolean.class, reservationId));
-    }
-
-    public Waiting findByReservationIdAndName(Long reservationId, String name) {
-        String sql = """
-            SELECT w.id AS id,
-                   w.name AS name,
-                   w.reservation_id AS reservation_id
-            FROM waiting w
-            WHERE w.reservation_id = ?
-              AND w.name = ?
-            """;
-        return jdbcTemplate.queryForObject(sql, waitingRowMapper, reservationId, name);
-    }
-
-    public void delete(Long id) {
-        jdbcTemplate.update("DELETE FROM waiting WHERE id = ?", id);
-    }
-}
