+++ src/main/java/roomescape/service/WaitingService.java
@@ -6,34 +6,34 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import roomescape.domain.Reservation;
-import roomescape.domain.Waiting;
+import roomescape.domain.ReservationSlot;
 import roomescape.dto.ReservationResponse;
 import roomescape.dto.WaitingRequest;
 import roomescape.dto.WaitingResponse;
 import roomescape.exception.CustomException;
 import roomescape.exception.ErrorCode;
+import roomescape.repository.ReservationSlotDao;
 import roomescape.repository.ReservationDao;
-import roomescape.repository.WaitingDao;
 
 @Service
 public class WaitingService {
-    private final WaitingDao waitingDao;
     private final ReservationDao reservationDao;
+    private final ReservationSlotDao reservationSlotDao;
 
-    public WaitingService(WaitingDao waitingDao, ReservationDao reservationDao) {
-        this.waitingDao = waitingDao;
+    public WaitingService(ReservationDao reservationDao, ReservationSlotDao reservationSlotDao) {
         this.reservationDao = reservationDao;
+        this.reservationSlotDao = reservationSlotDao;
     }
 
     public WaitingResponse save(LocalDateTime now, WaitingRequest request) {
-        Reservation reservation = reservationDao.findById(request.reservationId());
-        LocalDateTime time = LocalDateTime.of(reservation.getDate(), reservation.getTime().getStartAt());
+        ReservationSlot reservationSlot = reservationSlotDao.findById(request.reservationId());
+        LocalDateTime time = LocalDateTime.of(reservationSlot.getDate(), reservationSlot.getTime().getStartAt());
         validateDateAndTimeNotPast(now,time);
         try{
-            long waitingId = waitingDao.save(request.name(), request.reservationId());
-            int order =  waitingDao.findOrderByReservationId(waitingId, request.reservationId());
-            Waiting waiting = new Waiting(waitingId, request.name(), request.reservationId());
-            return WaitingResponse.from(waiting, order);
+            long waitingId = reservationDao.save(request.name(), request.reservationId());
+            int order =  reservationDao.findOrderByReservationId(waitingId, request.reservationId());
+            Reservation reservation = new Reservation(waitingId, request.name(), request.reservationId());
+            return WaitingResponse.from(reservation, order);
         } catch (DataIntegrityViolationException e) {
             throw new CustomException(ErrorCode.ALREADY_EXISTS_RESERVATION);
         }
@@ -41,18 +41,18 @@ public class WaitingService {
     }
 
     public void delete(LocalDateTime now, Long id) {
-        Waiting waiting = waitingDao.findById(id);
-        Reservation reservation = reservationDao.findById(waiting.getReservationId());
-        LocalDateTime time = LocalDateTime.of(reservation.getDate(), reservation.getTime().getStartAt());
+        Reservation reservation = reservationDao.findById(id);
+        ReservationSlot reservationSlot = reservationSlotDao.findById(reservation.getReservationId());
+        LocalDateTime time = LocalDateTime.of(reservationSlot.getDate(), reservationSlot.getTime().getStartAt());
         validateDateAndTimeNotPast(now, time);
 
-        waitingDao.delete(id);
+        reservationDao.delete(id);
     }
 
     public ReservationResponse findById(long id) {
-        Waiting waiting = waitingDao.findById(id);
-        Reservation reservation = reservationDao.findById(waiting.getReservationId());
-        return ReservationResponse.from(reservation, WaitingResponse.from(waiting, waitingDao.findOrderByReservationId(id, reservation.getId())));
+        Reservation reservation = reservationDao.findById(id);
+        ReservationSlot reservationSlot = reservationSlotDao.findById(reservation.getReservationId());
+        return ReservationResponse.from(reservationSlot, WaitingResponse.from(reservation, reservationDao.findOrderByReservationId(id, reservationSlot.getId())));
     }
     
     private void validateDateAndTimeNotPast(LocalDateTime now, LocalDateTime reservationTime) {
