+++ src/main/resources/data.sql
@@ -21,57 +21,57 @@ INSERT INTO theme (name, description, thumbnail_url) VALUES ('탐정 사무소',
 
 -- Reservations for popular theme ranking
 -- 공포의 저택 (theme 1) - 5 bookings
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-04-29', 3, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-06-30', 5, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-01', 7, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-02', 4, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-03', 8, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-04-29', 3, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-06-30', 5, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-01', 7, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-02', 4, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-03', 8, 1);
 
 -- 탐정 사무소 (theme 4) - 4 bookings
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-04-30', 6, 4);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-01', 9, 4);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-02', 11, 4);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-03', 3, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-04-30', 6, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-01', 9, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-02', 11, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-03', 3, 4);
 
 -- 마법사의 연구실 (theme 3) - 3 bookings
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-01', 2, 3);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-04', 6, 3);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-05', 10, 3);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-01', 2, 3);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-04', 6, 3);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-05', 10, 3);
 
 -- 우주 정거장 (theme 2) - 2 bookings
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-03', 4, 2);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-05', 8, 2);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-03', 4, 2);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-05', 8, 2);
 
 -- Future reservations after 2026-05-26
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-27', 3, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-28', 5, 2);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-29', 7, 4);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-30', 9, 3);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-05-31', 1, 1);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-06-01', 10, 2);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-06-02', 12, 3);
-INSERT INTO reservation (date, time_id, theme_id) VALUES ('2026-06-03', 13, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-27', 3, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-28', 5, 2);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-29', 7, 4);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-30', 9, 3);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-05-31', 1, 1);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-06-01', 10, 2);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-06-02', 12, 3);
+INSERT INTO reservation_slot (date, time_id, theme_id) VALUES ('2026-06-03', 13, 4);
 
 -- Waiting names linked to reservations
-INSERT INTO waiting (name, reservation_id) VALUES ('김철수', 1);
-INSERT INTO waiting (name, reservation_id) VALUES ('이영희', 2);
-INSERT INTO waiting (name, reservation_id) VALUES ('박민준', 3);
-INSERT INTO waiting (name, reservation_id) VALUES ('최수진', 4);
-INSERT INTO waiting (name, reservation_id) VALUES ('정다은', 5);
-INSERT INTO waiting (name, reservation_id) VALUES ('강현수', 6);
-INSERT INTO waiting (name, reservation_id) VALUES ('윤지원', 7);
-INSERT INTO waiting (name, reservation_id) VALUES ('임서준', 8);
-INSERT INTO waiting (name, reservation_id) VALUES ('한지아', 9);
-INSERT INTO waiting (name, reservation_id) VALUES ('송민재', 10);
-INSERT INTO waiting (name, reservation_id) VALUES ('오하린', 11);
-INSERT INTO waiting (name, reservation_id) VALUES ('문서윤', 12);
-INSERT INTO waiting (name, reservation_id) VALUES ('배준호', 13);
-INSERT INTO waiting (name, reservation_id) VALUES ('서지우', 14);
-INSERT INTO waiting (name, reservation_id) VALUES ('남도윤', 15);
-INSERT INTO waiting (name, reservation_id) VALUES ('권예린', 16);
-INSERT INTO waiting (name, reservation_id) VALUES ('조민성', 17);
-INSERT INTO waiting (name, reservation_id) VALUES ('안유진', 18);
-INSERT INTO waiting (name, reservation_id) VALUES ('신하늘', 19);
-INSERT INTO waiting (name, reservation_id) VALUES ('유지호', 20);
-INSERT INTO waiting (name, reservation_id) VALUES ('홍서연', 21);
-INSERT INTO waiting (name, reservation_id) VALUES ('백시우', 22);
+INSERT INTO reservation (name, reservation_id) VALUES ('김철수', 1);
+INSERT INTO reservation (name, reservation_id) VALUES ('이영희', 2);
+INSERT INTO reservation (name, reservation_id) VALUES ('박민준', 3);
+INSERT INTO reservation (name, reservation_id) VALUES ('최수진', 4);
+INSERT INTO reservation (name, reservation_id) VALUES ('정다은', 5);
+INSERT INTO reservation (name, reservation_id) VALUES ('강현수', 6);
+INSERT INTO reservation (name, reservation_id) VALUES ('윤지원', 7);
+INSERT INTO reservation (name, reservation_id) VALUES ('임서준', 8);
+INSERT INTO reservation (name, reservation_id) VALUES ('한지아', 9);
+INSERT INTO reservation (name, reservation_id) VALUES ('송민재', 10);
+INSERT INTO reservation (name, reservation_id) VALUES ('오하린', 11);
+INSERT INTO reservation (name, reservation_id) VALUES ('문서윤', 12);
+INSERT INTO reservation (name, reservation_id) VALUES ('배준호', 13);
+INSERT INTO reservation (name, reservation_id) VALUES ('서지우', 14);
+INSERT INTO reservation (name, reservation_id) VALUES ('남도윤', 15);
+INSERT INTO reservation (name, reservation_id) VALUES ('권예린', 16);
+INSERT INTO reservation (name, reservation_id) VALUES ('조민성', 17);
+INSERT INTO reservation (name, reservation_id) VALUES ('안유진', 18);
+INSERT INTO reservation (name, reservation_id) VALUES ('신하늘', 19);
+INSERT INTO reservation (name, reservation_id) VALUES ('유지호', 20);
+INSERT INTO reservation (name, reservation_id) VALUES ('홍서연', 21);
+INSERT INTO reservation (name, reservation_id) VALUES ('백시우', 22);
