+++ src/test/java/roomescape/controller/ReservationSlotControllerTest.java
@@ -19,7 +19,7 @@ import io.restassured.http.ContentType;
 
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationControllerTest {
+class ReservationSlotControllerTest {
     @Autowired
     private ReservationController reservationController;
