+++ src/test/java/roomescape/service/ReservationSlotServiceTest.java
@@ -22,7 +22,7 @@ import roomescape.exception.ErrorCode;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationServiceTest {
+class ReservationSlotServiceTest {
 
     @Autowired
     private ReservationService reservationService;
